Doc comments and cleanup
