+++ modules/src/ics04_channel/channel.rs
@@ -393,7 +393,7 @@ impl State {
         }
     }
 
-    // Parses the State out from a i32.
+    /// Parses the State out from a i32.
     pub fn from_i32(s: i32) -> Result<Self, Error> {
         match s {
             0 => Ok(Self::Uninitialized),
@@ -405,10 +405,20 @@ impl State {
         }
     }
 
+    /// Returns whether or not this channel state is `Open`.
     pub fn is_open(self) -> bool {
         self == State::Open
     }
 
+    /// Returns whether or not the channel with this state
+    /// has progressed less or the same than the argument.
+    ///
+    /// # Example
+    /// ```rust,ignore
+    /// assert!(State::Init.less_or_equal_progress(State::Open));
+    /// assert!(State::TryOpen.less_or_equal_progress(State::TryOpen));
+    /// assert!(!State::Closed.less_or_equal_progress(State::Open));
+    /// ```
     pub fn less_or_equal_progress(self, other: Self) -> bool {
         self as u32 <= other as u32
     }
