+++ relayer/src/chain/handle.rs
@@ -305,13 +305,17 @@ pub enum ChainRequest {
 dyn_clone::clone_trait_object!(ChainHandle);
 
 pub trait ChainHandle: DynClone + Send + Sync + Debug {
+    /// Get the [`ChainId`] of this chain.
     fn id(&self) -> ChainId;
 
+    /// Shutdown the chain runtime.
     fn shutdown(&self) -> Result<(), Error>;
 
+    /// Subscribe to the events emitted by the chain.
     fn subscribe(&self) -> Result<Subscription, Error>;
 
-    /// Send a transaction with `msgs` to chain.
+    /// Send the given `msgs` to the chain, packaged as one or more transactions,
+    /// and return the list of events emitted by the chain after the transaction was committed.
     fn send_msgs(&self, proto_msgs: Vec<prost_types::Any>) -> Result<Vec<IbcEvent>, Error>;
 
     fn get_signer(&self) -> Result<Signer, Error>;
